refactor(health,config): split oversized files by concern (pure code motion)
#2 file splits — no code changes, same packages, symbols identical; only
physical file boundaries move. goimports fixes per-file imports.

internal/health/checks.go (2987 lines) -> category files:
- checks_system.go, checks_incus.go, checks_network.go, checks_storage.go,
  checks_status.go, checks_probe.go, checks_monitoring.go
  (checks.go retains the shared package-level regex/var blocks.)

internal/config/config.go (1813 lines) -> concern files:
- types.go (config structs + accessor methods), defaults.go (GetDefaultConfig,
  profile synthesis, path/clone helpers), merge.go (Merge/ApplyProfile/
  inheritance + all merge*Into), validate.go (ProfileConfig.Validate)
  (config.go retains package-level consts + hostnameRe.)

Full build/vet/test green (23 pkgs); golangci-lint 0 issues; gofmt clean.
